pyproject.toml
[tool.poetry]
name = "ml_mobility_ns3"
version = "0.1.0"
description = ""
authors = ["racousin <racousin@student.42.fr>"]
readme = "README.md"
packages = [{include = "ml_mobility_ns3"}]

[tool.poetry.dependencies]
python = "^3.13"
torch = "^2.0.0"
pytorch-lightning = "^2.5.2"
hydra-core = "^1.3.2"
pandas = "^2.0.0"
numpy = "^1.24.0"
matplotlib = "^3.7.0"
seaborn = "^0.12.0"
jupyter = "^1.0.0"
tqdm = "^4.65.0"
geopandas = "^0.13.0"
folium = "^0.14.0"
scikit-learn = "^1.3.0"
tensorboard = "^2.13.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.0"
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"